Move CommandWrapper & ShellCommand to common
