+++ builder/amazon/chroot/builder.go
@@ -356,7 +356,7 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 	state.Put("awsSession", session)
 	state.Put("hook", hook)
 	state.Put("ui", ui)
-	state.Put("wrappedCommand", CommandWrapper(wrappedCommand))
+	state.Put("wrappedCommand", common.CommandWrapper(wrappedCommand))
 
 	// Build the steps
 	steps := []multistep.Step{
